Enable limiting the fuzz session to a given number of cycles (#7)
If every fuzz job that is defined in the configuration gets
executed once, that counts as one cycle. Controller.run API got
extended with a parameter that can be used to specify a limit for
the cycles, and the line-based UI also got extended with an
appropriate command line argument. (The TUI did not get the option
as it makes no sense there.)