Fix winrandom import error on Windows
A recent PyGithub version began to depend on pycrypto transitively,
which can easily fail with an `ImportError: No module named
'winrandom'` on Windows. This patch is a workaround to ignore the
error and postpone it to the point when `GithubReport` is actually
used. So, use the GitHub reporter on Windows only if you can
install pycrypto properly somehow.

(The issue was found by AppVeyor, which should go green now.)