Make sure that issues with different fields are handled correctly.
Based on a particular configuration, issue dictionaries can have
different fields even for the same SUT. However, until now,
at some places we had to explicitly refer to certain fields without
being prepared for that they may not exist. The patch gives fallbacks
to these places. The only functional change is that we use format
strings as bug report templates instead of the Template syntax. With
this change we can avoid KeyErrors when creating reports with
non-existent keys.