Support pattern-based input definition when listing a directory.
Until now, the ListDirectory class of Fuzzinator enabled
to go through all the files in a directory and test all of them
one-by-one. However, sometimes it's needed to filter them and only
focus on a subset (e.g. only test files with specific extension).
To support this need, the ``output` parameter of this class was
renamed to ``pattern`` and now it accepts shell or glob-like patterns.