Allow different Fuzzinator instances to use separate work directories
This changes allows the use of a {uid} placeholder in the path
of the working directory. The placeholder (if present) will be
replaced by the framework with the pid of the current Fuzzinator
instance. With this change, we can ensure that multiple Fuzzinator
instances won't interfere in their working directories.