Keep both the original test-case and the reduced one.
Until now, when a failing test-case got reduced, then the test-case
(that's the 'test' field of the issue dictionary) got replaced by
its reduced counterpart. However, in some use cases it's useful
to have access both to the original and the reduced version, too.
Hence, from now, the framework keeps the original test-case as it
was found in the 'test' field and saves the reduced version into
the 'reduced' field (which was a boolean field earlier denoting
whether the reduction has been executed already).