Signal load change event when the load has really changed
The `update_load` event is triggered on the listeners by the controller in
every case when the load is calculated. However, if the load did not change
there is no reason to call the `update_load` method to notify the listeners.

This modification will align the behaviour of the `update_load` invocation with
the intent defined in the `EventListener.update_load` method documentation.

The `EventListener.update_load` method doc states that it is
`Invoked when the framework's load changes`.