Move Trampoline out of the ListenerManager
The change extracts the `Trampoline` class from the `ListenerManager` class.

The main reason for the change is to improve the Windows support. During the
communication between processes the data gets pickled and the `ListenerManager`
also gets pickled. However as the `ListenerManager` has an internal class it
cannot be pickled, which results in an error during execution.

The reason for the `ListenerManager` getting pickled is that the `Controller`
object creates sub-processes and invokes its own methods thus all member of the
object get pickled including the `ListenerManager` instance.

Moving the `Trampoline` out of the manager will resolve this problem as the new
class is a top-level class which can be pickled.