Make ValidateJob a proper CallJob
When `CallJob`s `run`, they should return the issues newly found.
However until now, `ValidateJob`s returned an empty list if the
issue they validated couldn't be reproduced or a list with the
issue if reproduction was possible. Because of this violation of
the `CallJob` contract, `ValidateJob`s could not be treated the
same way as other `CallJob`s (which is the smaller problem) and
they could not send new issues, encountered during validation,
back to `Controller` for reduction (which is a more serious
problem).

This patch refactors `ValidateJob` by adding the new method
`validate` that can signal validity and new issues as separate
concepts while rewriting the `run` method as a `CallJob`-contract-
preserving wrapper around `validate` (i.e., it only returns the
list of new issues). Moreover, `ReduceJob` is also touched to make
use of the new `validate` functionality.