Fix conflicting -U command line argument definitions
The recent addition of the general `-U` command line argument (to
undefine a section/option of the configuration) introduced a
conflict with the TUI-specific `-U` argument (which forces UTF-8
encoding of widgets). To fix the conflict, this patch renames the
TUI-specific option to `--utf8` and `--utf-8`. (The patch adds two
alternative options names so that the presence or absence of the
dash doesn't have to be remembered.)