Update issues by their Object ID
Due to the existence of invalid issues, the sut/id pair isn't a
unique key anymore, hence it's not suitable to find the issues
to be updated. The patch switches to ObjectId-based update (which
is probably more efficient, too).
The patch also makes possible to use this method for invalidating issues.