Fix issue import
The add_issue method of MongoDriver is responsible for both importing
and adding new issues. However, the dictionary containing the issue
description is slightly different in these cases. The patch adds a
missing check to decide whether the source of the config is included
in the dictionary and handles it accordingly.