Add the concept of exporters
Exporters add a customization possibility by allowing the saving of
the issue in a SUT-specific, non-JSON, non-XML format. The custom
format doesn't have to contain all the details of the issue. In
fact, the simplest exporter only extracts the test case from the
issue (if the test case is a string or a byte array).