Fix the filter of current session queries
To filter for the issues and statistic of the current session,
the `first_seen` field is compared to the start time of the
active session. Since dates are stored in UTC format, we have to
use UTC format in the database query as well. This commit adapts
the `get_issues` and the `get_stats` method to this.