Save textual data as string instead of bytes
Apart from the `test` and `reduce` fields, every issue field is
saved as a string to help display information and avoid repeated
encoding to and decoding from bytes. The encodings can be defined in the
configuration, otherwise they are auto-detected. Since the `test`
and `reduce` fields contain critical data, i.e., the reproducing test
case, we keep them as bytes to be completely sure to have the original
test.