Pass only a single callback function to reducers
Reducers should not need (and should not access) other methods of
listeners but `on_job_progressed` only. Therefore, the reducer API
is changed to pass a bound partial method (the `on_job_progressed`
method of a listener object with `job_id` frozen) instead of a pair
of a listener and a job id.